+++ scheduler/secrets.go
@@ -3,15 +3,30 @@ package scheduler
 import (
 
 	// Kubernetes types
+	"strings"
+
 	log "github.com/sirupsen/logrus"
 	corev1 "k8s.io/api/core/v1"
 	apierrors "k8s.io/apimachinery/pkg/api/errors"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 )
 
+// syncSecret takes care of copying the primary TLS certificate from a configured
+// location into the lesson namespace. This is required because Kubernetes does not
+// allow cross-namespace secret lookups, and we need to be able to offer TLS for
+// http presentation endpoints.
 func (ls *LessonScheduler) syncSecret(nsName string) error {
 
-	prodCert, err := ls.Client.CoreV1().Secrets("prod").Get("tls-certificate", metav1.GetOptions{})
+	// Determine location of original certificate based from config
+	var certNs = "prod"
+	var certName = "tls-certificate"
+	certLocations := strings.Split(ls.SyringeConfig.CertLocation, "/")
+	if len(certLocations) == 2 {
+		certNs = certLocations[0]
+		certName = certLocations[1]
+	}
+
+	prodCert, err := ls.Client.CoreV1().Secrets(certNs).Get(certName, metav1.GetOptions{})
 	if err != nil {
 		return err
 	}