+++ CHANGELOG.md
@@ -2,12 +2,17 @@
 
 ## In development
 
+
+## v0.5.0 - February 01, 2020
+
 - Initial work on a 'mock' syringe with static sample data for integration testing [#136])https://github.com/nre-learning/syringe/pull/136)
 - Added cvx and frr image names to privileged container list in config.go [#129](https://github.com/nre-learning/syringe/pull/129)
 - Disable TSDB by default (configurable) [#130](https://github.com/nre-learning/syringe/pull/130)
 - Cleaned up and updated deps - installing grpc tooling from source [#135](https://github.com/nre-learning/syringe/pull/135)
 - Change SYRINGE_DOMAIN to optional variable, and provide default [#142](https://github.com/nre-learning/syringe/pull/142)
 - Add config option to control imagepullpolicy [#145](https://github.com/nre-learning/syringe/pull/145)
+- De-couple namespaces from tier - implement syringe ID option [#150](https://github.com/nre-learning/syringe/pull/150)
+- Move to DNS-based identifiers for http presentations [#151](https://github.com/nre-learning/syringe/pull/151)
 
 ## v0.4.0 - August 07, 2019